Rearrange code and refactor modules
