+++ .idea/modules.xml
@@ -2,9 +2,12 @@
 <project version="4">
   <component name="ProjectModuleManager">
     <modules>
+      <module fileurl="file://$PROJECT_DIR$/app/Android-android-architecture-example-app.iml" filepath="$PROJECT_DIR$/app/Android-android-architecture-example-app.iml" />
       <module fileurl="file://$PROJECT_DIR$/PokeApiMiniMVVM-PokeMiniMVVM.iml" filepath="$PROJECT_DIR$/PokeApiMiniMVVM-PokeMiniMVVM.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/PokeApiMiniMVVM-app.iml" filepath="$PROJECT_DIR$/app/PokeApiMiniMVVM-app.iml" />
       <module fileurl="file://$PROJECT_DIR$/PokeMiniMVVM.iml" filepath="$PROJECT_DIR$/PokeMiniMVVM.iml" />
+      <module fileurl="file://$PROJECT_DIR$/android-architecture-example.iml" filepath="$PROJECT_DIR$/android-architecture-example.iml" />
+      <module fileurl="file://$PROJECT_DIR$/app/android-architecture-example-app.iml" filepath="$PROJECT_DIR$/app/android-architecture-example-app.iml" />
       <module fileurl="file://$PROJECT_DIR$/app/app.iml" filepath="$PROJECT_DIR$/app/app.iml" />
     </modules>
   </component>
