+++ app/src/androidTest/kotlin/com/adriangl/pokeapi_mvvm/ui/pokemonlist/PokemonListActivityTest.kt
@@ -1,4 +1,4 @@
-package com.adriangl.pokeapi_mvvm.pokemonlist
+package com.adriangl.pokeapi_mvvm.ui.pokemonlist
 
 import android.content.Intent
 import android.view.View
