+++ app/src/main/kotlin/com/adriangl/pokeapi_mvvm/domain/pokemon/PokeStore.kt
@@ -1,13 +1,17 @@
-package com.adriangl.pokeapi_mvvm.pokemon
+package com.adriangl.pokeapi_mvvm.domain.pokemon
 
+import com.adriangl.pokeapi_mvvm.network.PokeApi
 import com.adriangl.pokeapi_mvvm.network.Pokemon
-import com.adriangl.pokeapi_mvvm.pokemonlist.FilterPokemonListAction
-import com.adriangl.pokeapi_mvvm.pokemonlist.GetPokemonDetailsListAction
-import com.adriangl.pokeapi_mvvm.pokemonlist.PokemonDetailsListLoadedAction
-import com.adriangl.pokeapi_mvvm.pokemonlist.PokemonListItem
+import com.adriangl.pokeapi_mvvm.ui.pokemonlist.PokemonListItem
 import mini.Reducer
 import mini.Store
 import mini.Task
+import mini.kodein.bindStore
+import org.kodein.di.Kodein
+import org.kodein.di.generic.bind
+import org.kodein.di.generic.instance
+import org.kodein.di.generic.singleton
+import retrofit2.Retrofit
 
 data class PokeState(
     val pokemonList: List<Pokemon>? = null,
@@ -41,5 +45,16 @@ class PokeStore(private val pokeController: PokeController) : Store<PokeState>()
                 filter = { pokemonListItem -> pokemonListItem.name.contains(action.query) }
             )
         )
+    }
+}
+
+object PokeModule {
+    fun create() = Kodein.Module("pokeStore") {
+        bindStore { PokeStore(instance()) }
+        bind<PokeController>() with singleton { PokeControllerImpl(instance(), instance()) }
+        bind<PokeApi>() with singleton {
+            val retrofit: Retrofit = instance()
+            retrofit.create(PokeApi::class.java)
+        }
     }
 }
