+++ app/src/main/kotlin/com/adriangl/pokeapi_mvvm/ui/pokemonlist/PokemonListActivity.kt
@@ -1,4 +1,4 @@
-package com.adriangl.pokeapi_mvvm.pokemonlist
+package com.adriangl.pokeapi_mvvm.ui.pokemonlist
 
 import android.os.Bundle
 import android.view.View
