+++ app/src/main/res/layout/pokemonlist_activity.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".pokemonlist.PokemonListActivity">
+    tools:context=".ui.pokemonlist.PokemonListActivity">
 
     <ProgressBar
         android:id="@+id/list_loading"
